+++ src/model/comments-model.js
@@ -1,5 +1,5 @@
 import { generateComment } from '../mock/comment';
-import { getRandomInteger } from '../utils';
+import { getRandomInteger } from '../utils/common';
 
 export default class CommentsModel {
   #comments = Array.from({ length: getRandomInteger(0, 10) }, generateComment);
